♻️#344 refactor: 모바일헤더 위 고정 및 NavSection 컴포넌트 분리
