+++ components/nav/Nav.tsx
@@ -27,7 +27,7 @@ const Nav = () => {
   }, [pathname]);
 
   return (
-    <nav aria-label='사이드바 네비게이션 메뉴'>
+    <nav aria-label='사이드바 네비게이션 메뉴' className='sticky top-0'>
       {/* 왼쪽 네브 (데스크탑, 테블릿) */}
       <motion.section
         initial='open'
